SphereCast and Code Organization 1
Implemented SphereCast instead of RayCast so that the laser can hit at a wider area. Also made some adjustments to the LaserDiode code like when the laser is flickering, before it turns back on, it uses the SphereCast before turning on the line renderer so that it doesn't show itself at the wrong position in a single frame.