Rough Draft: Maze and Follow Sight
Made a maze in the EnemyFollow_Alben scene to test the EnemyFollow agent can follow the player in a maze. Also, I made the rough draft for an EnemyFollowSight which uses a SphereCast to find the player to chase them. Once the player is behind cover, the enemy will go to the last place the player was scene. One issue for this is that if the player goes out of sight of the enemy like going behind it, the enemy will not know where the player is despite the player not being behind cover.