Code1
Got to 10:53 in the video. Since he uses a start button, I still need to make it so the player reaches a certain state in the level that triggers the dialogue.

Also, he uses a continue button. I need to change this to just pressing enter at some point.

Lastly, Unity is currently giving me an error regarding the Dialogue Trigger script. Will look into this next time.